Move modal close button top right
Move the close button of the modal to the top right
and simply display an X instead of the word "X".

This is mainly due to the fact that the modal grew in size
with the lecture details and thus closing always required the user to
scroll to the end. This is now simplified