Display location of each block
Display the blocks location next to the dates outside
the details view via the modal to offer the basic information
on first sight.

Add cypress UI tests for different viewports (mobile, desktop) to
ensure both display all required data. Somehow, I could not set the
cypress viewport beforeAll so that for each test I need to set it again
and re-visit the page which costs a lot of time.