Add time units to exam and workload durations
Add h for hours or min for minutes as the duration
for the exam as well as the workload in order to tell the
two apart.

If the exam does not have a duration, no unit is being displayed.

Split the integration tests into separate test files to test
them inndepedently and thus faster.

Fix the tests of the modal as the beforeEach 'modal' variable
was yielding always true test results. Now, the beforeEach
only opens the modal and then each it block finds its
sub-block and checks the value