Display hint if english name not present
Display a hint message if the english lecture name is not
present. This is the case of the lecture does not have
an entry in the module handbook.

In that case, the english name cannot be displayed in the lecture
list or the modal.

If the name is given, simply display it instead